HP-2482/Close_prices_and_notices_for_client_and_reseller
