+++ src/product/Application/BillingRegistryServiceInterface.php
@@ -13,7 +13,7 @@ use hiqdev\php\billing\product\quantity\FractionQuantityData;
 use hiqdev\php\billing\product\quantity\QuantityFormatterInterface;
 use hiqdev\php\billing\product\TariffTypeDefinitionInterface;
 
-interface BillingRegistryServiceInterface
+interface BillingRegistryServiceInterface extends BillingRegistryTariffServiseInterface, BillingRegistryBehaviorServiceInterface
 {
     /**
      * @param string $representationClass
@@ -23,32 +23,6 @@ interface BillingRegistryServiceInterface
 
     public function createQuantityFormatter(string $type, FractionQuantityData $data): QuantityFormatterInterface;
 
-    /**
-     * @param string $type - full type like 'overuse,lb_capacity_unit'
-     * @param string $behaviorClassWrapper
-     * @return BehaviorInterface
-     * @throws BehaviorNotFoundException
-     * @throws InvalidBehaviorException
-     */
-    public function getBehavior(string $type, string $behaviorClassWrapper): BehaviorInterface;
-
-    /**
-     * Find all behaviors attached to any TariffType or PriceType by specified Behavior class.
-     *
-     * @param string $behaviorClassWrapper
-     * @return Generator<BehaviorInterface>
-     */
-    public function getBehaviors(string $behaviorClassWrapper): Generator;
-
     public function getAggregate(string $type): AggregateInterface;
 
-    public function findTariffTypeDefinitionByBehavior(BehaviorInterface $behavior): TariffTypeDefinitionInterface;
-
-    /**
-     * Find all PriceTypeDefinition in registry by specified Behavior class.
-     *
-     * @param string $behaviorClassWrapper
-     * @return Generator<PriceTypeDefinitionInterface>
-     */
-    public function findPriceTypeDefinitionsByBehavior(string $behaviorClassWrapper): Generator;
 }
